+++ src/Exceptions/ValidationException.php
@@ -51,6 +51,9 @@ class ValidationException extends Exception implements ClientAware, ProvidesExte
         return true;
     }
 
+    /**
+     * @return array{validation: array<string, array<int, string>>}
+     */
     public function getExtensions(): array
     {
         return [