+++ src/Schema/AST/DocumentAST.php
@@ -178,9 +178,7 @@ class DocumentAST implements Serializable, Arrayable
 
         // Utilize the NodeList for lazy unserialization for performance gains.
         // Until they are accessed by name, they are kept in their array form.
-        // @phpstan-ignore-next-line TODO fixed in https://github.com/webonyx/graphql-php/pull/777
         $this->types = new NodeList($types);
-        // @phpstan-ignore-next-line TODO fixed in https://github.com/webonyx/graphql-php/pull/777
         $this->directives = new NodeList($directives);
     }
 }