+++ src/Schema/Directives/CanDirective.php
@@ -213,24 +213,16 @@ GRAPHQL;
         // should be [modelClassName, additionalArg, additionalArg...]
         array_unshift($arguments, $model);
 
-        // Gate responses were introduced in Laravel 6
-        // TODO remove with Laravel < 6 support
-        if (AppVersion::atLeast(6.0)) {
-            Utils::applyEach(
-                function ($ability) use ($gate, $arguments) {
-                    $response = $gate->inspect($ability, $arguments);
+        Utils::applyEach(
+            function ($ability) use ($gate, $arguments) {
+                $response = $gate->inspect($ability, $arguments);
 
-                    if ($response->denied()) {
-                        throw new AuthorizationException($response->message(), $response->code());
-                    }
-                },
-                $ability
-            );
-        } elseif (! $gate->check($ability, $arguments)) {
-            throw new AuthorizationException(
-                "You are not authorized to access {$this->nodeName()}"
-            );
-        }
+                if ($response->denied()) {
+                    throw new AuthorizationException($response->message(), $response->code());
+                }
+            },
+            $ability
+        );
     }
 
     /**