+++ src/Schema/Factories/FieldFactory.php
@@ -78,11 +78,7 @@ class FieldFactory
             ->send($fieldValue)
             ->through(array_merge($fieldMiddleware, $globalFieldMiddleware))
             ->via('handleField')
-            // TODO replace when we cut support for Laravel 5.6
-            //->thenReturn()
-            ->then(static function (FieldValue $fieldValue): FieldValue {
-                return $fieldValue;
-            })
+            ->thenReturn()
             ->getResolver();
 
         $fieldValue->setResolver(