+++ src/Subscriptions/Storage/CacheStorageManager.php
@@ -82,8 +82,7 @@ class CacheStorageManager implements StoresSubscriptions
         if ($this->ttl === null) {
             $this->cache->forever($channelKey, $subscriber);
         } else {
-            // TODO: Change to just pass the ttl directly when support for Laravel <=5.7 is dropped
-            $this->cache->put($channelKey, $subscriber, Carbon::now()->addSeconds($this->ttl));
+            $this->cache->put($channelKey, $subscriber, $this->ttl);
         }
     }
 
@@ -108,8 +107,7 @@ class CacheStorageManager implements StoresSubscriptions
         if ($this->ttl === null) {
             $this->cache->forever($key, $topic);
         } else {
-            // TODO: Change to just pass the ttl directly when support for Laravel <=5.7 is dropped
-            $this->cache->put($key, $topic, Carbon::now()->addSeconds($this->ttl));
+            $this->cache->put($key, $topic, $this->ttl);
         }
     }