+++ tests/Integration/Validation/ValidationTest.php
@@ -338,17 +338,13 @@ class ValidationTest extends TestCase
             ')
             ->assertGraphQLValidationError('bar', 'The bar must be at least 2 characters.');
 
-        $message = AppVersion::atLeast(8.32)
-            ? 'The bar must not be greater than 3 characters.'
-            : 'The bar may not be greater than 3 characters.';
-
         $this
             ->graphQL(/** @lang GraphQL */ '
             {
                 foo(bar: "fasdf")
             }
             ')
-            ->assertGraphQLValidationError('bar', $message);
+            ->assertGraphQLValidationError('bar', 'The bar must not be greater than 3 characters.');
     }
 
     public function testSingleFieldReferencesAreQualified(): void