+++ tests/Unit/Schema/Directives/ThrottleDirectiveTest.php
@@ -12,10 +12,6 @@ class ThrottleDirectiveTest extends TestCase
 {
     public function testNamedLimiter(): void
     {
-        if (AppVersion::below(8.0)) {
-            $this->markTestSkipped('Version less than 8.0 does not support named requests.');
-        }
-
         $this->schema = /** @lang GraphQL */'
         type Query {
             foo: Int @throttle(name: "test")
@@ -71,10 +67,6 @@ class ThrottleDirectiveTest extends TestCase
 
     public function testUnlimitedNamedLimiter(): void
     {
-        if (AppVersion::below(8.0)) {
-            $this->markTestSkipped('Version less than 8.0 does not support named requests.');
-        }
-
         $this->schema = /** @lang GraphQL */'
         type Query {
             foo: Int @throttle(name: "test")
@@ -82,12 +74,10 @@ class ThrottleDirectiveTest extends TestCase
         ';
 
         $rateLimiter = $this->createMock(RateLimiter::class);
-        // @phpstan-ignore-next-line phpstan ignores markTestSkipped
         $rateLimiter->expects(self::atLeast(1))
             ->method('limiter')
             ->with('test')
             ->willReturn(static function () {
-                // @phpstan-ignore-next-line phpstan ignores markTestSkipped
                 return Limit::none();
             });