+++ tests/Unit/Schema/Types/Scalars/DateScalarTest.php
@@ -48,11 +48,6 @@ abstract class DateScalarTest extends TestCase
 
     public function testConvertsCarbonCarbonImmutableToIlluminateSupportCarbon(): void
     {
-        // TODO remove when we stop supporting Laravel 5.7
-        if (! class_exists('\Carbon\CarbonImmutable')) {
-            $this->markTestSkipped('CarbonImmutable is not available with older Laravel versions');
-        }
-
         $this->assertInstanceOf(
             IlluminateCarbon::class,
             $this->scalarInstance()->parseValue(CarbonCarbonImmutable::now())